refactor: move `unknownErrorToString` to `shared`
