+++ packages/frontend/src/hooks/useProcessQr.ts
@@ -16,7 +16,7 @@ import type { T } from '@deltachat/jsonrpc-client'
 import type { WelcomeQrWithUrl } from '../contexts/InstantOnboardingContext'
 import type { TranslationKey } from '../../../shared/translationKeyType'
 import useChat from './chat/useChat'
-import { unknownErrorToString } from '../components/helpers/unknownErrorToString'
+import { unknownErrorToString } from '@deltachat-desktop/shared/unknownErrorToString'
 import ProxyConfiguration from '../components/dialogs/ProxyConfiguration'
 import { useSettingsStore } from '../stores/settings'
 import TransportsDialog from '../components/dialogs/Transports'
